Crate deeprust

DeepRust features:

The library contains the following functions:

The reg module contains functions to perform regularization operations

The metrics module contains functions to generate the confusion matrix

we use the confusion matrix macro to feed value into metric module functions

Modules

metrics
Module for benchmarking and analyzing the model's performance

reg
Module for performing regularization operations

Macros

Confusionmatrix
Confusionmatrix takes in four fields which are expresions which is converted into the confusion_matrix struct of the metrics module