Module deeprust::reg::regularizer

Functions inside these modules are used to penalize the update inorder to avoid over fitting and making a generalized module

Regularizer module contains the following modules:

L2 regularization

Functions

l2_reg
L2 Regularizer - Takes in lambda and theta as arguments for calculating the l2_reg value